renamed package session closed event
