+++ app/src/main/java/ru/evotor/framework/core/action/event/session/SessionClosedEvent.java
@@ -1,8 +1,8 @@
-package ru.evotor.framework.core.action.event.reports;
+package ru.evotor.framework.core.action.event.session;
 
 import android.support.annotation.Nullable;
 
-public class SessionClosedEvent extends ReportEvent {
+public class SessionClosedEvent extends SessionEvent {
     public static final String BROADCAST_ACTION_SESSION_CLOSED = "evotor.intent.action.reports.SESSION_CLOSED";
 
     public SessionClosedEvent() {
